Revert "Component Graphics (#40)"
This reverts commit 2e81205feb65622d85b0025ee441a3fb9c832d70.
