+++ src/components/Branch.java
@@ -3,30 +3,31 @@ package components;
 import static components.ComponentType.BRANCH;
 import static components.ComponentType.INPUT_PIN;
 import static components.ComponentType.OUTPUT_PIN;
+import static java.lang.Math.abs;
+import static java.lang.Math.min;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
+import java.awt.Color;
+import java.awt.Graphics;
+import java.awt.Point;
 
 import exceptions.InvalidIndexException;
 import exceptions.MalformedBranchException;
 import exceptions.MalformedGateException;
 
-/**
- * Corresponds to the {@link ComponentType#BRANCH BRANCH} type.
- *
- * @author alexm
- */
+/** Corresponds to the {@link ComponentType#BRANCH BRANCH} type */
 final class Branch extends Component {
 
-	private static final long serialVersionUID = 4L;
+	private static final long serialVersionUID = 3L;
 
 	private final Component	in, out;
 	private final int		indexIn, indexOut;
 
 	private boolean active = false;
 
-	private final ComponentGraphic g;
+	// graphics, 1 or -1, slope of the line
+	//  1 = draw top-left to bottom-right
+	// -1 = draw bottom-left to top-right
+	private int direction;
 
 	/**
 	 * Constructs a {@code Branch} between two Components at the specified indexes.
@@ -51,7 +52,6 @@ final class Branch extends Component {
 		if (outIndex >= outComponent.inCount())
 			throw new MalformedBranchException(outComponent, outIndex);
 
-		g = new BranchGraphics(this);
 		this.in = inComponent;
 		this.out = outComponent;
 		this.indexIn = inIndex;
@@ -79,10 +79,28 @@ final class Branch extends Component {
 		// repaint and propagate signal only if it's different
 		if (active != newActive) {
 			active = newActive;
-			getGraphics().repaint();
+			repaint();
 			out.wake_up(active, indexOut, hidden());
 		}
 	}
+
+	@Override
+	protected void destroySelf() {
+		checkChangeable();
+		in.removeOut(this, indexIn);
+		out.removeIn(this, indexOut);
+
+		// inform `out` that there is no longer an input
+		out.wake_up(false, indexOut);
+	}
+
+	@Override
+	protected void restoreDeletedSelf() {
+		connect();
+	}
+
+	@Override
+	protected void restoreSerialisedSelf() {}
 
 	@Override
 	protected boolean getActive(int index) {
@@ -110,45 +128,44 @@ final class Branch extends Component {
 		wake_up(in.getActive(indexIn));
 
 		// update the Branch's graphics
-		g.updateOnMovement();
+		updateOnMovement();
 	}
 
 	@Override
-	protected void destroySelf() {
-		checkChangeable();
-		in.removeOut(this, indexIn);
-		out.removeIn(this, indexOut);
-
-		// inform `out` that there is no longer an input
-		out.wake_up(false, indexOut);
+	protected void attachListeners() {
+		attachListeners_((byte) 0);
 	}
 
 	@Override
-	protected void restoreDeletedSelf() {
-		connect();
+	protected void draw(Graphics g) {
+		g.setColor(active ? Color.green : Color.red);
+
+		// draw with correct direction (as specified in the `direction` declaration)
+		if (direction == 1)
+			g.drawLine(5, 5, getWidth() - 5, getHeight() - 6);
+		else if (direction == -1)
+			g.drawLine(5, getHeight() - 6, getWidth() - 5, 5);
+		else
+			throw new RuntimeException("Invalid Branch direction");
+	}
+
+	/** @param g the Graphics object with which to draw the ID */
+	@Override
+	protected void drawID(Graphics g) {
+		g.setColor(Color.BLACK);
+		g.drawString(getID(), (getWidth() / 2) - 4, (getHeight() / 2) + 5);
 	}
 
 	@Override
-	protected void restoreSerialisedSelf() {}
-
-	@Override
-	protected List<Component> getInputs() {
-		List<Component> ls = new ArrayList<>();
-		ls.add(in);
-		return Collections.unmodifiableList(ls);
-	}
-
-	@Override
-	protected List<List<Component>> getOutputs() {
-		List<List<Component>> ls  = new ArrayList<>();
-		List<Component>       ls1 = new ArrayList<>();
-		ls1.add(out);
-		ls.add(Collections.unmodifiableList(ls1));
-		return Collections.unmodifiableList(ls);
-	}
-
-	@Override
-	public ComponentGraphic getGraphics() {
-		return g;
+	protected void updateOnMovement() {
+		// from the new coordinates calculate the Branch's start point, width and height
+		// and also calculate its direction (as specified in the declaration).
+		Point p1 = in.getBranchInputCoords(this, indexIn);
+		Point p2 = out.getBranchOutputCoords(this, indexOut);
+		direction = ((p2.x - p1.x) * (p2.y - p1.y)) > 0 ? 1 : -1;
+		// components with a dimension = 0 aren't drawn and text can't be drawn on a
+		// small space so add extra width/height here and remove it when drawing
+		setBounds(min(p1.x, p2.x) - 5, min(p1.y, p2.y) - 5, abs(p2.x - p1.x) + 11,
+				abs(p2.y - p1.y) + 11);
 	}
 }
