+++ src/components/GateAND.java
@@ -1,15 +1,36 @@
 package components;
 
-/**
- * A Primitive Gate that maps the inputs to their logical {@code and}.
- *
- * @author alexm
- */
+import java.awt.image.BufferedImage;
+import java.io.File;
+import java.io.IOException;
+
+import javax.imageio.ImageIO;
+
+import application.StringConstants;
+
+/** A Primitive Gate that maps the inputs to their logical {@code and}. */
 final class GateAND extends PrimitiveGate {
 
-	private static final long serialVersionUID = 4L;
+	private static final long serialVersionUID = 3L;
 
-	private final ComponentGraphic g;
+	private static final String sprite = StringConstants.COMPONENT_ICON_PATH
+	        + "gate_and.png";
+
+	private static final BufferedImage image;
+
+	static {
+		BufferedImage temp = null;
+		File file = null;
+
+		try {
+			file = new File(sprite);
+			temp = ImageIO.read(file);
+		} catch (IOException e) {
+			System.err.printf("Could not load image %s", file);
+		}
+
+		image = temp;
+	}
 
 	/**
 	 * Constructs the AND Gate with the given number of inputs and one output.
@@ -18,7 +39,6 @@ final class GateAND extends PrimitiveGate {
 	 */
 	GateAND(int in) {
 		super(in, 1);
-		g = new GateANDGraphics(this);
 	}
 
 	@Override
@@ -41,7 +61,7 @@ final class GateAND extends PrimitiveGate {
 	}
 
 	@Override
-	public ComponentGraphic getGraphics() {
-		return g;
+	protected BufferedImage getImage() {
+		return image;
 	}
 }
