+++ src/components/GateOR.java
@@ -1,15 +1,37 @@
 package components;
 
-/**
- * A Primitive Gate that maps the inputs to their logical {@code or}.
- *
- * @author alexm
- */
+import java.awt.image.BufferedImage;
+import java.io.File;
+import java.io.IOException;
+import java.util.function.Function;
+
+import javax.imageio.ImageIO;
+
+import application.StringConstants;
+
+/** A Primitive Gate that maps the inputs to their logical {@code or}. */
 final class GateOR extends PrimitiveGate {
 
-	private static final long serialVersionUID = 4L;
+	private static final long serialVersionUID = 2L;
 
-	private final ComponentGraphic g;
+	private static final String sprite = StringConstants.COMPONENT_ICON_PATH
+			+ "gate_or.png";
+
+	private static final BufferedImage image;
+
+	static {
+		BufferedImage temp = null;
+		File file = null;
+
+		try {
+			file = new File(sprite);
+			temp = ImageIO.read(file);
+		} catch (IOException e) {
+			System.err.printf("Could not load image %s", file);
+		}
+
+		image = temp;
+	}
 
 	/**
 	 * Constructs the OR Gate with the given number of inputs and one output.
@@ -18,7 +40,6 @@ final class GateOR extends PrimitiveGate {
 	 */
 	GateOR(int in) {
 		super(in, 1);
-		g = new GateORGraphic(this);
 	}
 
 	@Override
@@ -41,7 +62,12 @@ final class GateOR extends PrimitiveGate {
 	}
 
 	@Override
-	public ComponentGraphic getGraphics() {
-		return g;
+	protected BufferedImage getImage() {
+		return image;
+	}
+
+	@Override
+	protected Function<Integer, Integer> dxi() {
+		return i -> (int) (0.55 * Math.cos(1.2 * Math.asin(1 - (dyi().apply(i) / 20.0))) * 20);
 	}
 }
