+++ src/components/InputPin.java
@@ -2,31 +2,57 @@ package components;
 
 import static myUtil.Utility.foreach;
 
+import java.awt.Graphics;
+import java.awt.Point;
+import java.awt.image.BufferedImage;
+import java.io.File;
+import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 import java.util.Vector;
 
+import javax.imageio.ImageIO;
+
+import application.StringConstants;
 import exceptions.ComponentNotFoundException;
 import exceptions.MalformedGateException;
 
-/**
- * Corresponds to the {@link ComponentType#INPUT_PIN INPUT_PIN} type.
- *
- * @author alexm
- */
+/** Corresponds to the {@link ComponentType#INPUT_PIN INPUT_PIN} type. */
 final class InputPin extends Component {
 
-	private static final long serialVersionUID = 4L;
+	private static final long serialVersionUID = 3L;
 
-	private final Vector<Component> outputBranches;
+	private static final String sprite = StringConstants.COMPONENT_ICON_PATH
+			+ "input_pin_{state}.png";
+
+	private static final BufferedImage image_on, image_off;
+
+	static {
+		BufferedImage temp_on = null, temp_off = null;
+		File file = null;
+
+		try {
+			file = new File(sprite.replace("{state}", "on"));
+			temp_on = ImageIO.read(file);
+		} catch (IOException e) {
+			System.err.printf("Could not load image %s%n", file);
+		}
+
+		try {
+			file = new File(sprite.replace("{state}", "off"));
+			temp_off = ImageIO.read(file);
+		} catch (IOException e) {
+			System.err.printf("Could not load image %s%n", file);
+		}
+
+		image_on = temp_on;
+		image_off = temp_off;
+	}
+
+	private final Vector<Branch> outputBranches;
 	private boolean              active;
 
-	private final ComponentGraphic g;
-
-	/** Constructs an InputPin */
+	/** Constructs an {@code InputPin} */
 	InputPin() {
-		g = new InputPinGraphic(this);
 		outputBranches = new Vector<>(1, 1);
 		active = false;
 	}
@@ -34,6 +60,11 @@ final class InputPin extends Component {
 	@Override
 	public ComponentType type() {
 		return ComponentType.INPUT_PIN;
+	}
+
+	@Override
+	protected int inCount() {
+		return 0;
 	}
 
 	@Override
@@ -50,26 +81,22 @@ final class InputPin extends Component {
 		// propagate signal only if it's different
 		if (active != newActive) {
 			active = newActive;
-			getGraphics().repaint();
+			repaint();
 			foreach(outputBranches, b -> b.wake_up(active, hidden()));
 		}
 	}
 
 	@Override
-	protected int inCount() {
-		return 0;
+	protected void destroySelf() {
+		foreach(new ArrayList<>(outputBranches), Branch::destroy);
+		outputBranches.clear();
 	}
 
-	/**
-	 * Proper way for the client (the Factory) to set input. This method, unlike
-	 * wake_up, will throw when it is called on a hidden {@code InputPin}.
-	 *
-	 * @param newActive the new value for the active state of this InputPin
-	 */
-	void setActive(boolean newActive) {
-		checkChangeable();
-		wake_up(newActive);
-	}
+	@Override
+	protected void restoreDeletedSelf() {}
+
+	@Override
+	protected void restoreSerialisedSelf() {}
 
 	@Override
 	protected boolean getActive(int index) {
@@ -93,6 +120,17 @@ final class InputPin extends Component {
 	}
 
 	/**
+	 * Proper way for the client (the Factory) to set input. This method, unlike
+	 * wake_up, will throw when it is called on a hidden {@code InputPin}.
+	 *
+	 * @param newActive the new value for the active state of this InputPin
+	 */
+	void setActive(boolean newActive) {
+		checkChangeable();
+		wake_up(newActive);
+	}
+
+	/**
 	 * Marks this Component as unchangeable because it's hidden in a {@code Gate}.
 	 * Normally should only be called during the construction of a {@code Gate}.
 	 */
@@ -102,31 +140,28 @@ final class InputPin extends Component {
 	}
 
 	@Override
-	protected void destroySelf() {
-		foreach(getInputs(), Component::destroy);
-		outputBranches.clear();
+	protected void attachListeners() {
+		attachListeners_((byte) (DRAG_KB_FOCUS | ACTIVATE));
 	}
 
 	@Override
-	protected void restoreDeletedSelf() {}
-
-	@Override
-	protected void restoreSerialisedSelf() {}
-
-	@Override
-	protected List<Component> getInputs() {
-		return Collections.unmodifiableList(Collections.emptyList());
+	protected void draw(Graphics g) {
+		g.drawImage(getActive(0) ? image_on : image_off, 0, 0, null);
 	}
 
 	@Override
-	protected List<List<Component>> getOutputs() {
-		List<List<Component>> ls = new ArrayList<>();
-		ls.add(Collections.unmodifiableList(outputBranches));
-		return Collections.unmodifiableList(ls);
+	protected void updateOnMovement() {
+		// this component is moved by the user; tell branches to update
+		foreach(outputBranches, Branch::updateOnMovement);
 	}
 
 	@Override
-	public ComponentGraphic getGraphics() {
-		return g;
+	protected Point getBranchInputCoords(Branch b, int index) {
+		checkIndex(index, outCount());
+
+		if (!outputBranches.contains(b))
+			throw new ComponentNotFoundException(b, this);
+
+		return new Point((getX() + getWidth()) - 1, getY() + (getHeight() / 2));
 	}
 }
